added protected announcements route
