+++ api/src/main/java/com/codeforcommunity/api/IAnnouncementsProcessor.java
@@ -18,15 +18,6 @@ public interface IAnnouncementsProcessor {
   GetAnnouncementsResponse getAnnouncements(GetAnnouncementsRequest request);
 
   /**
-   * Creates a new announcement.
-   *
-   * @param request DTO containing the data for the announcement
-   * @param userData the JWT data for the user making the request
-   * @return the created announcement
-   */
-  PostAnnouncementResponse postAnnouncement(PostAnnouncementRequest request, JWTData userData);
-
-  /**
    * Gets all the announcements for a particular event.
    *
    * @param request DTO containing the event ID
@@ -34,23 +25,4 @@ public interface IAnnouncementsProcessor {
    */
   GetAnnouncementsResponse getEventSpecificAnnouncements(
       GetEventSpecificAnnouncementsRequest request);
-
-  /**
-   * Creates a new event-specific announcement.
-   *
-   * @param request DTO containing the data for the announcement
-   * @param userData the JWT data for the user making the request
-   * @param eventId the ID of the event
-   * @return the created announcement
-   */
-  PostAnnouncementResponse postEventSpecificAnnouncement(
-      PostAnnouncementRequest request, JWTData userData, int eventId);
-
-  /**
-   * Deletes an announcement.
-   *
-   * @param announcementId the ID of the announcement
-   * @param userData the JWT data for the user making the request
-   */
-  void deleteAnnouncement(int announcementId, JWTData userData);
 }
