+++ api/src/main/java/com/codeforcommunity/rest/subrouter/AnnouncementsRouter.java
@@ -3,12 +3,9 @@ package com.codeforcommunity.rest.subrouter;
 import static com.codeforcommunity.rest.ApiRouter.end;
 
 import com.codeforcommunity.api.IAnnouncementsProcessor;
-import com.codeforcommunity.auth.JWTData;
 import com.codeforcommunity.dto.announcements.GetAnnouncementsRequest;
 import com.codeforcommunity.dto.announcements.GetAnnouncementsResponse;
 import com.codeforcommunity.dto.announcements.GetEventSpecificAnnouncementsRequest;
-import com.codeforcommunity.dto.announcements.PostAnnouncementRequest;
-import com.codeforcommunity.dto.announcements.PostAnnouncementResponse;
 import com.codeforcommunity.rest.IRouter;
 import com.codeforcommunity.rest.RestFunctions;
 import io.vertx.core.Vertx;
@@ -34,10 +31,7 @@ public class AnnouncementsRouter implements IRouter {
     Router router = Router.router(vertx);
 
     registerGetAnnouncements(router);
-    registerPostAnnouncement(router);
     registerGetEventSpecificAnnouncements(router);
-    registerPostEventSpecificAnnouncement(router);
-    registerDeleteAnnouncement(router);
 
     return router;
   }
@@ -47,24 +41,9 @@ public class AnnouncementsRouter implements IRouter {
     getAnnouncementsRoute.handler(this::handleGetAnnouncements);
   }
 
-  private void registerPostAnnouncement(Router router) {
-    Route postAnnouncementRoute = router.post("/");
-    postAnnouncementRoute.handler(this::handlePostAnnouncement);
-  }
-
   private void registerGetEventSpecificAnnouncements(Router router) {
     Route getEventSpecificAnnouncementsRoute = router.get("/:event_id");
     getEventSpecificAnnouncementsRoute.handler(this::handleGetEventSpecificAnnouncements);
-  }
-
-  private void registerPostEventSpecificAnnouncement(Router router) {
-    Route postEventSpecificAnnouncementRoute = router.post("/:event_id");
-    postEventSpecificAnnouncementRoute.handler(this::handlePostEventSpecificAnnouncement);
-  }
-
-  private void registerDeleteAnnouncement(Router router) {
-    Route deleteAnnouncementRoute = router.delete("/:announcement_id");
-    deleteAnnouncementRoute.handler(this::handleDeleteAnnouncement);
   }
 
   private void handleGetAnnouncements(RoutingContext ctx) {
@@ -83,15 +62,6 @@ public class AnnouncementsRouter implements IRouter {
     end(ctx.response(), 200, JsonObject.mapFrom(response).toString());
   }
 
-  private void handlePostAnnouncement(RoutingContext ctx) {
-    PostAnnouncementRequest requestData =
-        RestFunctions.getJsonBodyAsClass(ctx, PostAnnouncementRequest.class);
-    JWTData userData = ctx.get("jwt_data");
-
-    PostAnnouncementResponse response = processor.postAnnouncement(requestData, userData);
-    end(ctx.response(), 200, JsonObject.mapFrom(response).toString());
-  }
-
   private void handleGetEventSpecificAnnouncements(RoutingContext ctx) {
     int eventId = RestFunctions.getRequestParameterAsInt(ctx.request(), "event_id");
     GetEventSpecificAnnouncementsRequest requestData =
@@ -99,26 +69,5 @@ public class AnnouncementsRouter implements IRouter {
 
     GetAnnouncementsResponse response = processor.getEventSpecificAnnouncements(requestData);
     end(ctx.response(), 200, JsonObject.mapFrom(response).encode());
-  }
-
-  private void handlePostEventSpecificAnnouncement(RoutingContext ctx) {
-    PostAnnouncementRequest requestData =
-        RestFunctions.getJsonBodyAsClass(ctx, PostAnnouncementRequest.class);
-    JWTData userData = ctx.get("jwt_data");
-
-    PostAnnouncementResponse response =
-        processor.postEventSpecificAnnouncement(
-            requestData,
-            userData,
-            RestFunctions.getRequestParameterAsInt(ctx.request(), "event_id"));
-    end(ctx.response(), 200, JsonObject.mapFrom(response).toString());
-  }
-
-  private void handleDeleteAnnouncement(RoutingContext ctx) {
-    int announcementId = RestFunctions.getRequestParameterAsInt(ctx.request(), "announcement_id");
-    JWTData userData = ctx.get("jwt_data");
-
-    processor.deleteAnnouncement(announcementId, userData);
-    end(ctx.response(), 200);
   }
 }
